feat: core: helpers: add define_ports decorator to help the definition of block ports
This removes the boiler plate code required to register ports in blocks
while mantains custom class attributes required to facilitate block
extensions and serialization
This commits adds a new method to the base block object called
init_ports(). This method is going to be called after the ports are
created and registered. Derived class can use this method to add
specific logic that needs to be executed after block ports creation
phase